requirements.txt
streamlit
google-generativeai
streamlit-lottie
python-dotenv
fitz
pyPDF2
setuptools
PyMuPDF
streamlit-authenticator
plotly
pandas